UCT/DC/GTEST: MP XRQ support for dc tl
