+++ src/uct/ib/dc/dc_mlx5.c
@@ -231,7 +231,7 @@ static unsigned uct_dc_mlx5_iface_progress(void *arg)
     uct_dc_mlx5_iface_t *iface = arg;
     unsigned count;
 
-    count = uct_rc_mlx5_iface_common_poll_rx(&iface->super, 0);
+    count = uct_rc_mlx5_iface_common_poll_rx(&iface->super, 0, 0);
     if (count > 0) {
         return count;
     }
@@ -243,7 +243,7 @@ static unsigned uct_dc_mlx5_iface_progress_tm(void *arg)
     uct_dc_mlx5_iface_t *iface = arg;
     unsigned count;
 
-    count = uct_rc_mlx5_iface_common_poll_rx(&iface->super, 1);
+    count = uct_rc_mlx5_iface_common_poll_rx(&iface->super, 1, 0);
     if (count > 0) {
         return count;
     }
