+++ src/uct/ib/dc/dc_mlx5_ep.c
@@ -645,15 +645,18 @@ ssize_t uct_dc_mlx5_ep_tag_eager_bcopy(uct_ep_h tl_ep, uct_tag_t tag,
 
     UCT_DC_MLX5_CHECK_RES(iface, ep);
 
-    UCT_RC_MLX5_FILL_TM_IMM(imm, app_ctx, ib_imm, opcode, MLX5_OPCODE_SEND, _IMM);
+    UCT_RC_MLX5_FILL_TM_IMM(imm, app_ctx, ib_imm, opcode, MLX5_OPCODE_SEND,
+                            _IMM);
 
     UCT_RC_MLX5_IFACE_GET_TM_BCOPY_DESC(&iface->super.super,
-                                        &iface->super.super.tx.mp, desc, tag,
-                                        app_ctx, pack_cb, arg, length);
+                                        iface->super.tm.bcopy_mp,
+                                        desc, tag, app_ctx, pack_cb,
+                                        arg, length);
 
     uct_dc_mlx5_iface_bcopy_post(iface, ep, opcode,
                                  sizeof(struct ibv_tmh) + length,
-                                 0, 0, desc, MLX5_WQE_CTRL_SOLICITED, ib_imm, desc + 1, NULL);
+                                 0, 0, desc, MLX5_WQE_CTRL_SOLICITED, ib_imm,
+                                 desc + 1, NULL);
 
     UCT_TL_EP_STAT_OP(&ep->super, TAG, BCOPY, length);
 
@@ -674,7 +677,8 @@ ucs_status_t uct_dc_mlx5_ep_tag_eager_zcopy(uct_ep_h tl_ep, uct_tag_t tag,
                        "uct_dc_mlx5_ep_tag_eager_zcopy");
     UCT_RC_CHECK_ZCOPY_DATA(sizeof(struct ibv_tmh),
                             uct_iov_total_length(iov, iovcnt),
-                            iface->super.super.super.config.seg_size);
+                            iface->super.tm.max_zcopy);
+
     UCT_DC_MLX5_CHECK_RES(iface, ep);
 
     UCT_RC_MLX5_FILL_TM_IMM(imm, app_ctx, ib_imm, opcode, MLX5_OPCODE_SEND, _IMM);
