+++ src/uct/ib/ud/accel/ud_mlx5.c
@@ -420,7 +420,7 @@ uct_ud_mlx5_iface_poll_rx(uct_ud_mlx5_iface_t *iface, int is_async)
 
     if (!uct_ud_iface_check_grh(&iface->super,
                                 UCS_PTR_BYTE_OFFSET(packet, UCT_IB_GRH_LEN),
-                                (ntohl(cqe->flags_rqpn) >> 28) & 3)) {
+                                uct_ib_mlx5_cqe_is_grh_present(cqe))) {
         ucs_mpool_put_inline(desc);
         goto out;
     }
