+++ test/gtest/uct/ib/test_rc.cc
@@ -312,317 +312,6 @@ UCS_TEST_P(test_rc_flow_control, fc_disabled_flush)
 UCT_INSTANTIATE_RC_TEST_CASE(test_rc_flow_control)
 
 
-#ifdef IBV_HW_TM
-// TODO: Remove when declared in UCT
-#define UCT_RC_MLX5_TAG_BCOPY_MAX     131072
-
-size_t test_rc_mp_xrq::m_rx_counter = 0;
-
-test_rc_mp_xrq::test_rc_mp_xrq() : m_hold_uct_desc(false),
-                                   m_first_received(false),
-                                   m_last_received(false)
-{
-    m_max_hdr        = sizeof(ibv_tmh) + sizeof(ibv_rvh);
-    m_uct_comp.count = 512; // We do not need completion func to be invoked
-    m_uct_comp.func  = NULL;
-}
-
-uct_rc_mlx5_iface_common_t* test_rc_mp_xrq::rc_mlx5_iface(entity &e)
-{
-    return ucs_derived_of(e.iface(), uct_rc_mlx5_iface_common_t);
-}
-
-void test_rc_mp_xrq::set_env_var_or_skip(void *config, const char *var,
-                                         const char *val)
-{
-    ucs_status_t status = uct_config_modify(config, var, val);
-    if (status != UCS_OK) {
-        ucs_warn("%s", ucs_status_string(status));
-        UCS_TEST_SKIP_R(std::string("Can't set ") + var);
-    }
-}
-
-void test_rc_mp_xrq::init()
-{
-    set_env_var_or_skip(m_iface_config, "RC_TM_NUM_STRIDES", "8");
-    set_env_var_or_skip(m_iface_config, "RC_TM_ENABLE", "y");
-    set_env_var_or_skip(m_md_config, "MLX5_DEVX_OBJECTS", "dct,dcsrq,rcsrq,rcqp");
-
-    uct_test::init();
-
-    uct_iface_params params;
-    params.field_mask  = UCT_IFACE_PARAM_FIELD_RX_HEADROOM     |
-                         UCT_IFACE_PARAM_FIELD_OPEN_MODE       |
-                         UCT_IFACE_PARAM_FIELD_HW_TM_EAGER_CB  |
-                         UCT_IFACE_PARAM_FIELD_HW_TM_EAGER_ARG |
-                         UCT_IFACE_PARAM_FIELD_HW_TM_RNDV_CB   |
-                         UCT_IFACE_PARAM_FIELD_HW_TM_RNDV_ARG;
-
-    // tl and dev names are taken from resources via GetParam, no need
-    // to fill it here
-    params.rx_headroom = 0;
-    params.open_mode   = UCT_IFACE_OPEN_MODE_DEVICE;
-    params.eager_cb    = unexp_eager;
-    params.eager_arg   = reinterpret_cast<void*>(this);
-    params.rndv_cb     = unexp_rndv;
-    params.rndv_arg    = reinterpret_cast<void*>(this);
-
-    entity *sender = uct_test::create_entity(params);
-    m_entities.push_back(sender);
-
-    entity *receiver = uct_test::create_entity(params);
-    m_entities.push_back(receiver);
-
-    if (!UCT_RC_MLX5_MP_ENABLED(rc_mlx5_iface(test_rc_mp_xrq::sender()))) {
-        UCS_TEST_SKIP_R("No MP XRQ support");
-    }
-
-    sender->connect(0, *receiver, 0);
-
-    uct_iface_set_am_handler(receiver->iface(), AM_ID, am_handler, this, 0);
-}
-
-void test_rc_mp_xrq::send_eager_bcopy(mapped_buffer *buf)
-{
-    ssize_t len = uct_ep_tag_eager_bcopy(sender().ep(0), 0x11,
-                                         reinterpret_cast<uint64_t>(this),
-                                         mapped_buffer::pack,
-                                         reinterpret_cast<void*>(buf), 0);
-
-    EXPECT_EQ(buf->length(), static_cast<size_t>(len));
-}
-
-void test_rc_mp_xrq::send_eager_zcopy(mapped_buffer *buf)
-{
-    UCS_TEST_GET_BUFFER_IOV(iov, iovcnt, buf->ptr(), buf->length(), buf->memh(),
-                            sender().iface_attr().cap.tag.eager.max_iov);
-
-    ucs_status_t status = uct_ep_tag_eager_zcopy(sender().ep(0), 0x11,
-                                                 reinterpret_cast<uint64_t>(this),
-                                                 iov, iovcnt, 0, &m_uct_comp);
-    ASSERT_UCS_OK_OR_INPROGRESS(status);
-}
-
-void test_rc_mp_xrq::send_rndv_zcopy(mapped_buffer *buf)
-{
-    UCS_TEST_GET_BUFFER_IOV(iov, iovcnt, buf->ptr(), buf->length(), buf->memh(),
-                            sender().iface_attr().cap.tag.rndv.max_iov);
-
-    uint64_t dummy_hdr       = 0xFAFA;
-    ucs_status_ptr_t rndv_op = uct_ep_tag_rndv_zcopy(sender().ep(0), 0x11, &dummy_hdr,
-                                                     sizeof(dummy_hdr), iov,
-                                                     iovcnt, 0, &m_uct_comp);
-    ASSERT_FALSE(UCS_PTR_IS_ERR(rndv_op));
-
-    // There will be no real RNDV performed, cancel the op to avoid mpool
-    // warning on exit
-    ASSERT_UCS_OK(uct_ep_tag_rndv_cancel(sender().ep(0),rndv_op));
-}
-
-void test_rc_mp_xrq::send_rndv_request(mapped_buffer *buf)
-{
-    size_t size = sender().iface_attr().cap.tag.rndv.max_hdr;
-    void *hdr   = alloca(size);
-
-    ASSERT_UCS_OK(uct_ep_tag_rndv_request(sender().ep(0), 0x11, hdr, size, 0));
-}
-
-void test_rc_mp_xrq::send_am_bcopy(mapped_buffer *buf)
-{
-    ssize_t len = uct_ep_am_bcopy(sender().ep(0), AM_ID, mapped_buffer::pack,
-                                  reinterpret_cast<void*>(buf), 0);
-
-    EXPECT_EQ(buf->length(), static_cast<size_t>(len));
-}
-
-void test_rc_mp_xrq::test_common(send_func sfunc, size_t num_segs,
-                                 size_t exp_segs, bool is_eager)
-{
-    size_t seg_size  = rc_mlx5_iface(sender())->super.super.config.seg_size;
-    size_t seg_num   = is_eager ? num_segs : 1;
-    size_t exp_val   = is_eager ? exp_segs : 1;
-    size_t size      = (seg_size * seg_num) - m_max_hdr;
-    m_rx_counter     = 0;
-    m_first_received = m_last_received = false;
-
-    EXPECT_TRUE(size <= sender().iface_attr().cap.tag.eager.max_bcopy);
-    mapped_buffer buf(size, SEND_SEED, sender());
-
-    (this->*sfunc)(&buf);
-
-    wait_for_value(&m_rx_counter, exp_val, true);
-    EXPECT_EQ(exp_val, m_rx_counter);
-    EXPECT_EQ(is_eager, m_first_received); // relevant for eager only
-    EXPECT_EQ(is_eager, m_last_received);  // relevant for eager only
-}
-
-ucs_status_t test_rc_mp_xrq::handle_uct_desc(void *data, unsigned flags)
-{
-    if ((flags & UCT_CB_PARAM_FLAG_DESC) && m_hold_uct_desc) {
-        m_uct_descs.push_back(data);
-        return UCS_INPROGRESS;
-    }
-
-    return UCS_OK;
-}
-
-ucs_status_t test_rc_mp_xrq::am_handler(void *arg, void *data, size_t length,
-                                        unsigned flags)
-{
-   EXPECT_TRUE(flags & UCT_CB_PARAM_FLAG_FIRST);
-   EXPECT_FALSE(flags & UCT_CB_PARAM_FLAG_MORE);
-
-   m_rx_counter++;
-
-   test_rc_mp_xrq *self = reinterpret_cast<test_rc_mp_xrq*>(arg);
-   return self->handle_uct_desc(data, flags);
-}
-
-ucs_status_t test_rc_mp_xrq::unexp_handler(void *data, unsigned flags,
-                                           uint64_t imm, void **context)
-{
-    void *self = reinterpret_cast<void*>(this);
-
-    if (flags & UCT_CB_PARAM_FLAG_FIRST) {
-        // Set the message context which will be passed back with the rest of
-        // message fragments
-        *context         = self;
-        m_first_received = true;
-
-    } else {
-        // Check that the correct message context is passed with all fragments
-        EXPECT_EQ(self, *context);
-    }
-
-    if (!(flags & UCT_CB_PARAM_FLAG_MORE)) {
-        // Last message should contain valid immediate value
-        EXPECT_EQ(reinterpret_cast<uint64_t>(this), imm);
-        m_last_received = true;
-    } else {
-        // Immediate value is passed with the last message only
-        EXPECT_EQ(0ul, imm);
-    }
-
-
-    return handle_uct_desc(data, flags);
-}
-
-ucs_status_t test_rc_mp_xrq::unexp_eager(void *arg, void *data, size_t length,
-                                         unsigned flags, uct_tag_t stag,
-                                         uint64_t imm, void **context)
-{
-    test_rc_mp_xrq *self = reinterpret_cast<test_rc_mp_xrq*>(arg);
-
-    m_rx_counter++;
-
-    return self->unexp_handler(data, flags, imm, context);
-}
-
-ucs_status_t test_rc_mp_xrq::unexp_rndv(void *arg, unsigned flags,
-                                        uint64_t stag, const void *header,
-                                        unsigned header_length,
-                                        uint64_t remote_addr, size_t length,
-                                        const void *rkey_buf)
-{
-    EXPECT_FALSE(flags & UCT_CB_PARAM_FLAG_FIRST);
-    EXPECT_FALSE(flags & UCT_CB_PARAM_FLAG_MORE);
-
-    m_rx_counter++;
-
-    return UCS_OK;
-}
-
-UCS_TEST_P(test_rc_mp_xrq, config)
-{
-    uct_rc_mlx5_iface_common_t *iface = rc_mlx5_iface(sender());
-
-    // MP XRQ is supported with tag offload only
-    EXPECT_TRUE(UCT_RC_MLX5_TM_ENABLED(iface));
-
-    // With MP XRQ segment size should be equal to MTU, because HW generates
-    // CQE per each received MTU
-    size_t mtu = uct_ib_mtu_value(uct_ib_iface_port_attr(&(iface)->super.super)->active_mtu);
-    EXPECT_EQ(mtu, iface->super.super.config.seg_size);
-
-    const uct_iface_attr *attrs = &sender().iface_attr();
-
-    // Max tag bcopy is limited by tag tx memory pool
-    EXPECT_EQ(iface->tm.max_bcopy - sizeof(ibv_tmh),
-              attrs->cap.tag.eager.max_bcopy);
-    EXPECT_GT(attrs->cap.tag.eager.max_bcopy,
-              iface->super.super.config.seg_size);
-
-    // Max tag zcopy is limited by maximal IB message size
-    EXPECT_EQ(uct_ib_iface_port_attr(&iface->super.super)->max_msg_sz - sizeof(ibv_tmh),
-              attrs->cap.tag.eager.max_zcopy);
-
-    // Maximal AM size should not exceed segment size, so it would always
-    // arrive in one-fragment packet (with header it should be strictly less)
-    EXPECT_LT(attrs->cap.am.max_bcopy, iface->super.super.config.seg_size);
-    EXPECT_LT(attrs->cap.am.max_zcopy, iface->super.super.config.seg_size);
-}
-
-UCS_TEST_P(test_rc_mp_xrq, desc_release)
-{
-    m_hold_uct_desc = true; // We want to "hold" UCT memory descriptors
-    std::pair<send_func, bool> sfuncs[5] = {
-              std::make_pair(&test_rc_mp_xrq::send_eager_bcopy,  true),
-              std::make_pair(&test_rc_mp_xrq::send_eager_zcopy,  true),
-              std::make_pair(&test_rc_mp_xrq::send_rndv_zcopy,   false),
-              std::make_pair(&test_rc_mp_xrq::send_rndv_request, false),
-              std::make_pair(&test_rc_mp_xrq::send_am_bcopy,     false)
-    };
-
-    for (int i = 0; i < 5; ++i) {
-        test_common(sfuncs[i].first, 3, 3, sfuncs[i].second);
-    }
-
-    for (ucs::ptr_vector<void>::const_iterator iter = m_uct_descs.begin();
-         iter != m_uct_descs.end(); ++iter)
-    {
-        uct_iface_release_desc(*iter);
-    }
-}
-
-UCS_TEST_P(test_rc_mp_xrq, am)
-{
-    test_common(&test_rc_mp_xrq::send_am_bcopy, 1, 1, false);
-}
-
-UCS_TEST_P(test_rc_mp_xrq, bcopy_eager_only)
-{
-    test_common(&test_rc_mp_xrq::send_eager_bcopy, 1);
-}
-
-UCS_TEST_P(test_rc_mp_xrq, zcopy_eager_only)
-{
-    test_common(&test_rc_mp_xrq::send_eager_zcopy, 1);
-}
-
-UCS_TEST_P(test_rc_mp_xrq, bcopy_eager)
-{
-    test_common(&test_rc_mp_xrq::send_eager_bcopy, 5, 5);
-}
-
-UCS_TEST_P(test_rc_mp_xrq, zcopy_eager)
-{
-    test_common(&test_rc_mp_xrq::send_eager_zcopy, 5, 5);
-}
-
-UCS_TEST_P(test_rc_mp_xrq, rndv_zcopy)
-{
-    test_common(&test_rc_mp_xrq::send_rndv_zcopy, 1, 1, false);
-}
-
-UCS_TEST_P(test_rc_mp_xrq, rndv_request)
-{
-    test_common(&test_rc_mp_xrq::send_rndv_request, 1, 1, false);
-}
-
-_UCT_INSTANTIATE_TEST_CASE(test_rc_mp_xrq, rc_mlx5)
-#endif
-
-
 #if ENABLE_STATS
 
 void test_rc_flow_control_stats::test_general(int wnd, int soft_thresh,
